\name{ggdistr}
\alias{ggdistr}
\title{Plotting GGSDT distributions}
\usage{
ggdistr(mu2 = 1.5, alpha2 = 1.2, beta = 1.8)
}
\arguments{
  \item{mu2}{mean value for S2 distribution}
  \item{alpha2}{alpha value for S2 distribution}
  \item{beta}{beta value for S1 and S2 distributions}
}
\description{
Plots GGSDT internal distributions with ggplot2
}
\examples{
ggdistr(mu2 = 1.5, alpha2 = 1.2, beta = 1.8)
}
