\name{ggroc1}
\alias{ggroc1}
\title{Plotting GGSDT type1 ROC}
\usage{
ggroc1(mu2 = 1.5, alpha2 = 1.2, beta = 1.8)
}
\arguments{
  \item{mu2}{mean value for S2 distribution}
  \item{alpha2}{alpha value for S2 distribution}
  \item{beta}{beta value for S1 and S2 distributions}
}
\description{
Plots GGSDT type1 ROC with ggplot2
}
\examples{
ggroc1(mu2 = 1.5, alpha2 = 1.2, beta = 1.8)
}
