\name{ggzroc1}
\alias{ggzroc1}
\title{Plotting GGSDT type1 zROC}
\usage{
ggzroc1(mu2 = 1.5, alpha2 = 1.2, beta = 1.8)
}
\arguments{
  \item{mu2}{mean value for S2 distribution}
  \item{alpha2}{alpha value for S2 distribution}
  \item{beta}{beta value for S1 and S2 distributions}
}
\description{
Plots GGSDT type1 ROC in z-transformed space with ggplot2
}
\examples{
ggzroc1(mu2 = 1.5, alpha2 = 1.2, beta = 1.8)
}
